add clicks and some logic to DetailsScreen

+++ app/src/main/java/com/example/coroutinescomposeapp/model/DescriptionItemUI.kt
@@ -6,7 +6,7 @@ data class DescriptionItemUI(
     @DrawableRes val itemImages: List<Int>,
     val itemName: String,
     val itemDescription: String,
-    val itemPrice: Int,
+    val itemPrice: Double,
     val itemRating: Double,
     val itemCountOfReviews: Int,
     val colors: List<String>,
